Move search result transform into lib
