+++ packages/gitbook/src/lib/search.ts
